them theme, vote, chinh lai db cho phu hop (file export)

+++ SocialNetworkBE/src/main/java/com/tlqt/services/PostService.java
@@ -26,5 +26,5 @@ public interface PostService {
 
     List<Post> getPostsByAuthorId(int authorId);
     
-    int countPostsByUserId(int userId);
+    long countPostsByUserId(int userId);
 }
